+++ apps/webpack.js
@@ -307,141 +307,6 @@ function devtool(options) {
   }
 }
 
-// Customize webpack config for storybook.
-// @param {Object} sbConfig - Webpack configuration from storybook library.
-function storybookConfig(sbConfig) {
-  return {
-    ...sbConfig,
-    // Overwrite aliases
-    resolve: {
-      ...sbConfig.resolve,
-      ...baseConfig.resolve,
-      alias: {
-        ...baseConfig.resolve.alias,
-        '@cdo/apps/lib/util/firehose': path.resolve(__dirname, 'test', 'util'),
-      },
-    },
-    // Overwrite rules
-    module: {
-      ...sbConfig.module,
-      ...baseConfig.module,
-      rules: baseConfig.module.rules,
-    },
-    // Overwrite externals
-    externals: {
-      blockly: 'this Blockly',
-    },
-    // Extend plugins
-    plugins: [
-      ...sbConfig.plugins,
-      new webpack.ProvidePlugin({React: 'react'}),
-      new webpack.DefinePlugin({
-        IN_UNIT_TEST: JSON.stringify(false),
-        IN_STORYBOOK: JSON.stringify(true),
-        'process.env.mocha_entry': JSON.stringify(process.env.mocha_entry),
-        'process.env.NODE_ENV': JSON.stringify(
-          envConstants.NODE_ENV || 'development'
-        ),
-        PISKEL_DEVELOPMENT_MODE: JSON.stringify(false),
-      }),
-    ],
-  };
-}
-
-// config for our test runner
-var karmaConfig = _.extend({}, baseConfig, {
-  devtool: devtool(),
-  resolve: _.extend({}, baseConfig.resolve, {
-    alias: _.extend({}, baseConfig.resolve.alias, {
-      '@cdo/locale': path.resolve(
-        __dirname,
-        'test',
-        'util',
-        'locale-do-not-import.js'
-      ),
-      '@cdo/netsim/locale': path.resolve(
-        __dirname,
-        'test',
-        'util',
-        'netsim',
-        'locale-do-not-import.js'
-      ),
-      '@cdo/applab/locale': path.resolve(
-        __dirname,
-        'test',
-        'util',
-        'applab',
-        'locale-do-not-import.js'
-      ),
-      '@cdo/gamelab/locale': path.resolve(
-        __dirname,
-        'test',
-        'util',
-        'gamelab',
-        'locale-do-not-import.js'
-      ),
-      '@cdo/music/locale': path.resolve(
-        __dirname,
-        'test',
-        'util',
-        'music',
-        'locale-do-not-import.js'
-      ),
-      '@cdo/javalab/locale': path.resolve(
-        __dirname,
-        'test',
-        'util',
-        'javalab',
-        'locale-do-not-import.js'
-      ),
-      '@cdo/weblab/locale': path.resolve(
-        __dirname,
-        'test',
-        'util',
-        'weblab',
-        'locale-do-not-import.js'
-      ),
-      '@cdo/tutorialExplorer/locale': path.resolve(
-        __dirname,
-        'test',
-        'util',
-        'tutorialExplorer',
-        'locale-do-not-import.js'
-      ),
-      firebase: path.resolve(__dirname, 'test', 'util', 'MockFirebase.js'),
-      // Use mock-firmata to unit test playground-io maker components
-      firmata: 'mock-firmata/mock-firmata',
-    }),
-  }),
-  externals: {
-    blockly: 'this Blockly',
-
-    // The below are necessary for enzyme to work.
-    // See https://github.com/airbnb/enzyme/blob/master/docs/guides/webpack.md
-    cheerio: 'window',
-    'react/addons': true,
-    'react/lib/ExecutionEnvironment': true,
-    'react/lib/ReactContext': true,
-  },
-  plugins: [
-    new webpack.ProvidePlugin({
-      React: 'react',
-      Buffer: ['buffer', 'Buffer'],
-      process: 'process/browser',
-    }),
-    new webpack.DefinePlugin({
-      IN_UNIT_TEST: JSON.stringify(true),
-      IN_STORYBOOK: JSON.stringify(false),
-      'process.env.mocha_entry': JSON.stringify(process.env.mocha_entry),
-      'process.env.NODE_ENV': JSON.stringify(
-        envConstants.NODE_ENV || 'development'
-      ),
-      LEVEL_TYPE: JSON.stringify(envConstants.LEVEL_TYPE),
-      PISKEL_DEVELOPMENT_MODE: JSON.stringify(false),
-    }),
-  ],
-});
-
 /**
  * Generate the appropriate webpack config based off of our base config and
  * some input options
@@ -515,8 +380,7 @@ function create(options) {
 }
 
 module.exports = {
-  config: baseConfig,
-  karmaConfig: karmaConfig,
-  storybookConfig: storybookConfig,
-  create: create,
+  baseConfig,
+  devtool,
+  create,
 };
